[sparkle] Move the shared side-panel model into components/
stories/conversationSidePanels.tsx was shared code living in the stories
directory. Merge it into components/ConversationSidePanels.tsx, which
already held the fake Files and Credit usage panels, and repoint Inbox,
Pods and People_Agent at the new location so stories/ holds only stories.
