+++ sparkle/playground/src/stories/Inbox.tsx
@@ -78,6 +78,16 @@ import {
 } from "react";
 
 import { AgentBuilderView } from "../components/AgentBuilderView";
+import {
+  ConversationActions,
+  conversationFilesFor,
+  isFileView,
+  type SelectedCitation,
+  type SidePanelView,
+  sidePanelContent,
+  sidePanelLabel,
+  sidePanelSizing,
+} from "../components/ConversationSidePanels";
 import { ConversationView } from "../components/ConversationView";
 import { CreateRoomDialog } from "../components/CreateRoomDialog";
 import { GroupConversationView } from "../components/GroupConversationView";
@@ -134,16 +144,6 @@ import {
   resolvePodContext,
   shouldShowMemberChrome,
 } from "./podPanelConfig";
-import {
-  ConversationActions,
-  conversationFilesFor,
-  isFileView,
-  type SelectedCitation,
-  type SidePanelView,
-  sidePanelContent,
-  sidePanelLabel,
-  sidePanelSizing,
-} from "./conversationSidePanels";
 import TemplateSelection, { type Template } from "./TemplateSelection";
 
 // ── Types ─────────────────────────────────────────────────────────────────────
